+++ analyses/fusion-filtering/run_fusion_merged.sh
@@ -2,7 +2,7 @@
 
 # K S Gaonkar
 
-# Run fusion_filtering
+# Run fusion-filtering
 
 set -e
 set -o pipefail
@@ -22,29 +22,29 @@ polya_expression_file="data/pbta-gene-expression-rsem-fpkm.polya.rds"
 stranded_expression_file="data/pbta-gene-expression-rsem-fpkm.stranded.rds"
 
 # directory that holds all of the reference files
-references_directory="analyses/fusion_filtering/references/"
+references_directory="analyses/fusion-filtering/references/"
 normal_expression_file="$references_directory/Brain_FPKM_hg38_matrix.txt.zip"
 
 # data release files to use for recurrent fusion/fused genes detection
 # TODO: once the data release gets updated to v15+, use the putative oncogenic fusion file included in the data download
-putative_oncogenic_fusion="analyses/fusion_filtering/results/pbta-fusion-putative-oncogenic.tsv"
+putative_oncogenic_fusion="analyses/fusion-filtering/results/pbta-fusion-putative-oncogenic.tsv"
 histologies_file="data/pbta-histologies.tsv"
 independent_samples_file="data/independent-specimens.wgswxs.primary-plus.tsv"
 
 # results folder for analysis
-results_folder="analyses/fusion_filtering/results/"
+results_folder="analyses/fusion-filtering/results/"
 
 
 # Run Fusion standardization
-Rscript analyses/fusion_filtering/01-fusion-standardization.R --fusionfile $arriba_file \
+Rscript analyses/fusion-filtering/01-fusion-standardization.R --fusionfile $arriba_file \
                                                               --caller "arriba" \
                                                               --outputfile $standard_arriba_file
-Rscript analyses/fusion_filtering/01-fusion-standardization.R --fusionfile $starfusion_file \
+Rscript analyses/fusion-filtering/01-fusion-standardization.R --fusionfile $starfusion_file \
                                                               --caller "starfusion" \
                                                               --outputfile $standard_starfusion_file
 
 # Run Fusion general filtering for polya
-Rscript analyses/fusion_filtering/02-fusion-filtering.R --standardFusionFiles $standard_starfusion_file,$standard_arriba_file  \
+Rscript analyses/fusion-filtering/02-fusion-filtering.R --standardFusionFiles $standard_starfusion_file,$standard_arriba_file  \
                                                         --expressionMatrix $polya_expression_file \
                                                         --artifactFilter $artifact_filter  \
                                                         --readingFrameFilter $reading_frame_filter \
@@ -52,7 +52,7 @@ Rscript analyses/fusion_filtering/02-fusion-filtering.R --standardFusionFiles $s
                                                         --outputfile scratch/standardFusionPolyaExp \
                                                         --readthroughFilter
 # Run Fusion general filtering for stranded
-Rscript analyses/fusion_filtering/02-fusion-filtering.R --standardFusionFiles $standard_arriba_file,$standard_starfusion_file \
+Rscript analyses/fusion-filtering/02-fusion-filtering.R --standardFusionFiles $standard_arriba_file,$standard_starfusion_file \
                                                         --expressionMatrix $stranded_expression_file \
                                                         --artifactFilter $artifact_filter \
                                                         --readingFrameFilter $reading_frame_filter \
@@ -62,24 +62,24 @@ Rscript analyses/fusion_filtering/02-fusion-filtering.R --standardFusionFiles $s
 
 
 # Fusion zscore annotation for filtered fusion
-Rscript analyses/fusion_filtering/03-Calc-zscore-annotate.R --standardFusionCalls scratch/standardFusionPolyaExp_QC_expression_filtered_annotated.RDS \
+Rscript analyses/fusion-filtering/03-Calc-zscore-annotate.R --standardFusionCalls scratch/standardFusionPolyaExp_QC_expression_filtered_annotated.RDS \
                                                             --expressionMatrix $polya_expression_file \
                                                             --normalExpressionMatrix $normal_expression_file \
                                                             --outputfile scratch/standardFusionPolyaExp_QC_expression
 
-Rscript analyses/fusion_filtering/03-Calc-zscore-annotate.R --standardFusionCalls scratch/standardFusionStrandedExp_QC_expression_filtered_annotated.RDS \
+Rscript analyses/fusion-filtering/03-Calc-zscore-annotate.R --standardFusionCalls scratch/standardFusionStrandedExp_QC_expression_filtered_annotated.RDS \
                                                             --expressionMatrix $stranded_expression_file \
                                                             --normalExpressionMatrix $normal_expression_file \
                                                             --outputfile scratch/standardFusionStrandedExp_QC_expression
 
 # Project specific filtering
-Rscript -e "rmarkdown::render('analyses/fusion_filtering/04-project-specific-filtering.Rmd')"
+Rscript -e "rmarkdown::render('analyses/fusion-filtering/04-project-specific-filtering.Rmd')"
 
 # QC filter putative oncogene found in more than 4 histologies
-Rscript -e "rmarkdown::render('analyses/fusion_filtering/05-QC_putative_onco_fusion_dustribution.Rmd')"
+Rscript -e "rmarkdown::render('analyses/fusion-filtering/05-QC_putative_onco_fusion_dustribution.Rmd')"
 
 # Recurrent fusion/fused genes
-Rscript analyses/fusion_filtering/06-recurrent-fusions-per-histology.R --standardFusionCalls $putative_oncogenic_fusion \
+Rscript analyses/fusion-filtering/06-recurrent-fusions-per-histology.R --standardFusionCalls $putative_oncogenic_fusion \
                                                                        --clinicalFile $histologies_file \
                                                                        --outputfolder $results_folder \
                                                                        --independentSpecimensFile $independent_samples_file