+++ analyses/tp53-nf1-score/run_classifier.sh
@@ -19,7 +19,7 @@ scratch_dir="scratch"
 cds_file="${scratch_dir}/gencode.v27.primary_assembly.annotation.bed"
 consensus_file="${data_dir}/pbta-snv-consensus-mutation.maf.tsv.gz"
 clinical_file="${data_dir}/pbta-histologies.tsv"
-analysis_dir="analyses/tp53_nf1_score"
+analysis_dir="analyses/tp53-nf1-score"
 
 # Convert GTF to BED file
 # Here we are only extracting lines with as a CDS i.e. are coded in protein